feat: Add append functionality to file transfer client
This commit adds the ability to append a new line to the received file in the file transfer client. It introduces a 3-second delay using the `time.sleep()` function before appending the new line. This feature enhances the functionality of the file transfer client by allowing the client to modify the received file.